[coro_rpc] user don't need define router
