--- include/coro_rpc/coro_rpc/protocol/router_impl.hpp
@@ -1,268 +1,0 @@
-/*
- * Copyright (c) 2022, Alibaba Group Holding Limited;
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-#pragma once
-#include <functional>
-#include <memory>
-#include <string>
-#include <string_view>
-#include <unordered_map>
-#include <variant>
-
-#include "../connection.hpp"
-#include "../coro_connection.hpp"
-#include "../rpc_execute.hpp"
-#include "easylog/easylog.h"
-#include "router.hpp"
-#include "struct_pack/struct_pack/md5_constexpr.hpp"
-#include "util/function_name.h"
-#include "util/type_traits.h"
-namespace coro_rpc::protocol {
-inline auto router::get_handler(uint32_t id)
-    -> std::function<std::optional<std::string>(
-        std::string_view, rpc_conn, rpc_protocol::req_header &req_head,
-        typename rpc_protocol::supported_serialize_protocols protocols)> * {
-  if (auto it = handlers_.find(id); it != handlers_.end()) {
-    return &it->second;
-  }
-  return nullptr;
-}
-
-inline auto router::get_coro_handler(uint32_t id)
-    -> std::function<async_simple::coro::Lazy<std::optional<std::string>>(
-        std::string_view, rpc_conn, rpc_protocol::req_header &req_head,
-        typename rpc_protocol::supported_serialize_protocols protocols)> * {
-  if (auto it = coro_handlers_.find(id); it != coro_handlers_.end()) {
-    return &it->second;
-  }
-  return nullptr;
-}
-
-inline async_simple::coro::Lazy<std::pair<std::errc, std::string>>
-router::route_coro(
-    auto handler, std::string_view data, rpc_conn conn,
-    rpc_protocol::req_header &header,
-    typename rpc_protocol::supported_serialize_protocols protocols) {
-  using namespace std::string_literals;
-  if (handler) [[likely]] {
-    try {
-#ifndef NDEBUG
-      if (auto it = id2name_.find(header.function_id); it != id2name_.end()) {
-        ELOGV(INFO, "route coro function name %s", it->second.data());
-      }
-#endif
-      // clang-format off
-      auto res = co_await (*handler)(data, std::move(conn), header, protocols);
-      // clang-format on
-      if (res.has_value()) [[likely]] {
-        co_return std::make_pair(std::errc{}, std::move(res.value()));
-      }
-      else {  // deserialize failed
-        ELOGV(ERROR, "the rpc payload deserialize failed, function id %d",
-              header.function_id);
-        co_return std::make_pair(std::errc::invalid_argument,
-                                 "invalid rpc function arguments"s);
-      }
-    } catch (const std::exception &e) {
-      ELOGV(ERROR, "the rpc function has exception %s, function id %d",
-            e.what(), header.function_id);
-      co_return std::make_pair(std::errc::interrupted, e.what());
-    } catch (...) {
-      ELOGV(ERROR, "the rpc function has unknown exception, function id %d",
-            header.function_id);
-      co_return std::make_pair(std::errc::interrupted, "unknown exception"s);
-    }
-  }
-  else [[unlikely]] {
-    ELOGV(ERROR, "the rpc function not found, function id %d",
-          header.function_id);
-    co_return std::make_pair(std::errc::function_not_supported,
-                             "the function not found"s);
-  }
-}
-
-inline std::pair<std::errc, std::string> router::route(
-    auto handler, std::string_view data, rpc_conn conn,
-    rpc_protocol::req_header &header,
-    typename rpc_protocol::supported_serialize_protocols protocols) {
-  using namespace std::string_literals;
-  if (handler) [[likely]] {
-    try {
-#ifndef NDEBUG
-      if (auto it = id2name_.find(header.function_id); it != id2name_.end()) {
-        ELOGV(INFO, "route function name %s", it->second.data());
-      }
-#endif
-      auto res = (*handler)(data, std::move(conn), header, protocols);
-      if (res.has_value()) [[likely]] {
-        return std::make_pair(std::errc{}, std::move(res.value()));
-      }
-      else {  // deserialize failed
-        ELOGV(ERROR, "the rpc payload deserialize failed, function id %d",
-              header.function_id);
-        return std::make_pair(std::errc::invalid_argument,
-                              "invalid rpc function arguments"s);
-      }
-    } catch (const std::exception &e) {
-      ELOGV(ERROR, "the rpc function has exception %s, function id %d",
-            e.what(), header.function_id);
-      return std::make_pair(std::errc::interrupted, e.what());
-    } catch (...) {
-      ELOGV(ERROR, "the rpc function has unknown exception, function id %d",
-            header.function_id);
-      return std::make_pair(std::errc::interrupted,
-                            "unknown rpc function exception"s);
-    }
-  }
-  else [[unlikely]] {
-    ELOGV(ERROR, "the rpc function not found, function id %d",
-          header.function_id);
-    return std::make_pair(std::errc::function_not_supported,
-                          "the function not found"s);
-  }
-}
-// See https://gcc.gnu.org/bugzilla/show_bug.cgi?id=100611
-// We use this struct instead of lambda for workaround
-template <typename rpc_protocol, auto Func, typename Self>
-struct execute_visitor {
-  std::string_view data;
-  rpc_conn conn;
-  typename rpc_protocol::req_header &req_head;
-  Self *self;
-  template <typename serialize_protocol>
-  async_simple::coro::Lazy<std::optional<std::string>> operator()(
-      const serialize_protocol &) {
-    return internal::execute_coro<rpc_protocol, serialize_protocol, Func>(
-        data, std::move(conn), req_head, self);
-  }
-};
-
-template <typename rpc_protocol, auto Func>
-struct execute_visitor<rpc_protocol, Func, void> {
-  std::string_view data;
-  rpc_conn conn;
-  typename rpc_protocol::req_header &req_head;
-  template <typename serialize_protocol>
-  async_simple::coro::Lazy<std::optional<std::string>> operator()(
-      const serialize_protocol &) {
-    return internal::execute_coro<rpc_protocol, serialize_protocol, Func>(
-        data, std::move(conn), req_head);
-  }
-};
-
-template <auto func, typename Self>
-inline void router::regist_one_handler(Self *self) {
-  if (self == nullptr) [[unlikely]] {
-    ELOGV(CRITICAL, "null connection!");
-  }
-
-  constexpr auto name = get_func_name<func>();
-  constexpr auto id =
-      struct_pack::MD5::MD5Hash32Constexpr(name.data(), name.length());
-  using return_type = function_return_type_t<decltype(func)>;
-  if constexpr (is_specialization_v<return_type, async_simple::coro::Lazy>) {
-    auto it = coro_handlers_.emplace(
-        id,
-        [self](std::string_view data, rpc_conn conn,
-               typename rpc_protocol::req_header &req_head,
-               typename rpc_protocol::supported_serialize_protocols protocols) {
-          execute_visitor<rpc_protocol, func, Self> visitor{
-              data, std::move(conn), req_head, self};
-          return std::visit(visitor, protocols);
-        });
-    if (!it.second) {
-      ELOGV(CRITICAL, "duplication function %s register!", name.data());
-    }
-  }
-  else {
-    auto it = handlers_.emplace(
-        id,
-        [self](std::string_view data, rpc_conn conn,
-               typename rpc_protocol::req_header &req_head,
-               typename rpc_protocol::supported_serialize_protocols protocols) {
-          return std::visit(
-              [data, conn = std::move(conn), self,
-               &req_head]<typename serialize_protocol>(
-                  const serialize_protocol &obj) mutable {
-                return internal::execute<rpc_protocol, serialize_protocol,
-                                         func>(data, std::move(conn), req_head,
-                                               self);
-              },
-              protocols);
-        });
-    if (!it.second) {
-      ELOGV(CRITICAL, "duplication function %s register!", name.data());
-    }
-  }
-
-  id2name_.emplace(id, name);
-}
-
-template <auto func>
-inline void router::regist_one_handler() {
-  static_assert(!std::is_member_function_pointer_v<decltype(func)>,
-                "register member function but lack of the parent object");
-  using return_type = function_return_type_t<decltype(func)>;
-
-  constexpr auto name = get_func_name<func>();
-  constexpr auto id =
-      struct_pack::MD5::MD5Hash32Constexpr(name.data(), name.length());
-  if constexpr (is_specialization_v<return_type, async_simple::coro::Lazy>) {
-    auto it = coro_handlers_.emplace(
-        id, [](std::string_view data, rpc_conn conn,
-               typename rpc_protocol::req_header &req_head,
-               typename rpc_protocol::supported_serialize_protocols protocols) {
-          execute_visitor<rpc_protocol, func, void> visitor{
-              data, std::move(conn), req_head};
-          return std::visit(visitor, protocols);
-        });
-    if (!it.second) {
-      ELOGV(CRITICAL, "duplication function %s register!", name.data());
-    }
-  }
-  else {
-    auto it = handlers_.emplace(
-        id, [](std::string_view data, rpc_conn conn,
-               typename rpc_protocol::req_header &req_head,
-               typename rpc_protocol::supported_serialize_protocols protocols) {
-          return std::visit(
-              [data, conn = std::move(conn),
-               &req_head]<typename serialize_protocol>(
-                  const serialize_protocol &obj) {
-                return internal::execute<rpc_protocol, serialize_protocol,
-                                         func>(data, std::move(conn), req_head);
-              },
-              protocols);
-        });
-    if (!it.second) {
-      ELOGV(CRITICAL, "duplication function %s register!", name.data());
-    }
-  }
-  id2name_.emplace(id, name);
-}
-
-template <auto first, auto... func>
-inline void router::regist_handler(class_type_t<decltype(first)> *self) {
-  regist_one_handler<first>(self);
-  (regist_one_handler<func>(self), ...);
-}
-
-template <auto first, auto... func>
-inline void router::regist_handler() {
-  regist_one_handler<first>();
-  (regist_one_handler<func>(), ...);
-}
-
-}  // namespace coro_rpc::protocol
