+++ include/coro_rpc/rpc_connection.hpp
@@ -17,4 +17,3 @@
 
 #include <coro_rpc/coro_rpc/connection.hpp>
 #include <coro_rpc/coro_rpc/coro_connection.hpp>
-#include <coro_rpc/coro_rpc/protocol/router_impl.hpp>
